fix: chapter02 mission02 피어 리뷰 후 수정
