+++ mission/chapter02/todoList/src/types/index.tsx
@@ -1,0 +1,4 @@
+export type Task = {
+  id: number;
+  text: string;
+};
